Add triggers to .github/workflows/ (#52)
* Add workflow_dispatch trigger to workflows from GitHub

* Update GitHub Actions checkout version

This commit updates the version of the `actions/checkout` action
in the following workflows:

- `.github/workflows/android.yml`
- `.github/workflows/copy-branch.yml`

The version is updated from `v2` to `v4`.

* Update java distribution to zulu

This commit updates the java distribution to zulu in the
`.github/workflows/android.yml` workflow.

* Update JDK version to 17

This commit updates the JDK version used in the
`.github/workflows/android.yml` workflow.

- Updates the `java-version` from 11 to 17.
- Adds gradle cache.